Revert "Revert "refactor: rename file to accurately reflect RFC 327 PR number""
This reverts commit 0f8eb5716234aba7bd279eb56cf234e4ef726a08.
