+++ peril.settings.json
@@ -33,10 +33,10 @@
       "pull_request": "dangerfile.ts"
     },
     "artsy/volt": {
-      "pull_request": "dangerfile.ts"
+      "pull_request": ["dangerfile.ts"]
     },
     "artsy/peril-settings": {
-      "pull_request": ["dangerfile.ts", "org/rfc_237.ts"]
+      "pull_request": ["dangerfile.ts", "org/rfc_327.ts"]
       // "pull_request.review_requested": "org/addReviewer.ts"
     }
   },
